extract helpers for LoadStackNodes
